Started working on Zombie

+++ Assets/Scripts/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-	void TakeDamage(float dmg);
+	void TakeDamage(int dmg);
 }
